reuse threadfunction to publish shm host
