+++ rgbd/src/rgbd_to_shm.cpp
@@ -3,7 +3,6 @@
 #include <ros/master.h>
 #include <ros/names.h>
 #include <ros/node_handle.h>
-#include <ros/publisher.h>
 #include <ros/rate.h>
 
 #include "rgbd/client_rgbd.h"
@@ -14,22 +13,9 @@
 
 #include <std_msgs/String.h>
 
+#include <functional>
 #include <thread>
 
-// ----------------------------------------------------------------------------------------
-
-void pubHostnameThreadFunc(ros::NodeHandle& nh, const std::string server_name, const std::string hostname, const float frequency)
-{
-    ros::Publisher pub_shm_hostname = nh.advertise<std_msgs::String>(server_name + "/hosts", 1);
-    ros::Rate r(frequency);
-    std_msgs::String msg;
-    msg.data = hostname;
-    while(nh.ok())
-    {
-        pub_shm_hostname.publish(msg);
-        r.sleep();
-    }
-}
 
 // ----------------------------------------------------------------------------------------
 
@@ -54,7 +40,7 @@ int main(int argc, char **argv)
     server.initialize(server_name);
 
     ros::NodeHandle nh;
-    std::thread pub_hostname_thread(pubHostnameThreadFunc, std::ref(nh), server_name, host_name, 10);
+    std::thread pub_hostname_thread(rgbd::pubHostnameThreadFunc, std::ref(nh), server_name, host_name, 10);
 
     rgbd::ImagePtr image_ptr;
