+++ rgbd/src/server_shm.cpp
@@ -6,6 +6,11 @@
 #include <boost/interprocess/sync/scoped_lock.hpp>
 
 #include <ros/console.h>
+#include <ros/node_handle.h>
+#include <ros/publisher.h>
+#include <ros/rate.h>
+
+#include <std_msgs/String.h>
 
 namespace ipc = boost::interprocess;
 
@@ -110,6 +115,21 @@ void ServerSHM::send(const Image& image)
 
         buffer_header_->cond_empty.notify_one();
         ++buffer_header_->sequence_nr;
+    }
+}
+
+// ----------------------------------------------------------------------------------------
+
+void pubHostnameThreadFunc(ros::NodeHandle& nh, const std::string server_name, const std::string hostname, const float frequency)
+{
+    ros::Publisher pub_shm_hostname = nh.advertise<std_msgs::String>(server_name + "/hosts", 1);
+    ros::Rate r(frequency);
+    std_msgs::String msg;
+    msg.data = hostname;
+    while(nh.ok())
+    {
+        pub_shm_hostname.publish(msg);
+        r.sleep();
     }
 }
